Cell/Virus Classes and Board Class framework
Implemented Cell and Virus classes fully.  Implemented some basic frame
working for board class, including drawing the cells and virus, and
animating the virus.  ISSUES: Cannot get the thread to run for the
board class.  Not sure why, but the run method is just never getting
called for board and so cycle nor animate is ever getting called.