Game Over Panel
Set up game over panel; however, we do not know how to transition from
the actual game screen to the game over panel. Also, we are trying to
set up either a cell count or virus count but we can’t get the variable
to decrement the right way or where to call the decrement method.