Body Cell Image Attempt
We uploaded an image of a body cell to be overlaid on top of the body
cell rectangles. However, when we went to draw them on top of the
rectangles, we ran into errors. So, we commented the changes out so we
can work on them later. We also changed the initial color of the
rectangles to green.