JTextPane added to sidebar, glitch remains
Added the JTextPane to sidebar.  Even before this was added there was a
paint issue where the scoreboard and background of the game screen would
be painted to the sidebar every time the sidebar was
repainted/revalidated. I suspect it's arising from somewhere in the
Board class. Given that this occurs with the JTextArea as well, I think
it has to do with the sidebarPanel as a whole. Kept all the old code in
case we want to undo this.